+++ bin/cli.php
@@ -32,6 +32,7 @@ use App\Console\MigrateIdActionIndex;
 use App\Console\MigrateMailLogTokens;
 use App\Console\MigrateIpCreatedDayIndex;
 use App\Console\MigrateDropMailLogColumns;
+use App\Console\MigrateDropMailLogIndexes;
 
 use App\Console\MigrateDb;
 use App\Console\OptimizeTable;
@@ -56,6 +57,7 @@ $application->add(new MigrateIdActionIndex());
 $application->add(new MigrateMailLogTokens());
 $application->add(new MigrateIpCreatedDayIndex());
 $application->add(new MigrateDropMailLogColumns());
+$application->add(new MigrateDropMailLogIndexes());
 $application->add(new MigrateDb());
 
 $application->run();